more review changes ... WIP
